// src-ui/kernel/types.ts
export interface AnyObject {
  [k: string]: any;
}

// 返回封装
export interface Response {
  code: number;
  msg: string;
  data: AnyObject;
}
// ...
export interface TauriApp {
  getName: Function;
  getTauriVersion: Function;
  getVersion: Function;
  show: Function;
  hide: Function;
}

// 继承TauriCli
export interface TauriCli {
  getMatches: Function;
}

// 继承TauriClipboard
export interface TauriClipboard {
  writeText: Function;
  readText: Function;
}

// 继承TauriDialog
export interface TauriDialog {
  open: Function;
  save: Function;
  message: Function;
  ask: Function;
  confirm: Function;
}

// 继承TauriEvent
export interface TauriEvent {
  listen: Function;
  once: Function;
  emit: Function;
}

// 继承TauriFs
export interface TauriFs {
  readTextFile: Function;
  readBinaryFile: Function;
  writeTextFile: Function;
  writeBinaryFile: Function;
  readDir: Function;
  createDir: Function;
  removeDir: Function;
  copyFile: Function;
  removeFile: Function;
  renameFile: Function;
  exists: Function;
}

// 继承TauriGlobalShortcut
export interface TauriGlobalShortcut {
  register: Function;
  registerAll: Function;
  isRegistered: Function;
  unregister: Function;
  unregisterAll: Function;
}

// 继承TauriHttp
export interface TauriHttp {
  Body: AnyObject;
  Client: AnyObject;
  Response: AnyObject;
  ResponseType: AnyObject;
  fetch: AnyObject;
  getClient: AnyObject;
}

// 继承TauriNotification
export interface TauriNotification {
  isPermissionGranted: Function;
  requestPermission: Function;
  sendNotification: Function;
}

// 继承TauriPath
export interface TauriPath {
  appDir: Function;
  appConfigDir: Function;
  appDataDir: Function;
  appLocalDataDir: Function;
  appCacheDir: Function;
  audioDir: Function;
  cacheDir: Function;
  configDir: Function;
  dataDir: Function;
  desktopDir: Function;
  documentDir: Function;
  downloadDir: Function;
  executableDir: Function;
  fontDir: Function;
  homeDir: Function;
  localDataDir: Function;
  pictureDir: Function;
  publicDir: Function;
  resourceDir: Function;
  resolveResource: Function;
  runtimeDir: Function;
  templateDir: Function;
  videoDir: Function;
  logDir: Function;
  appLogDir: Function;
  sep: string;
  delimiter: string;
  resolve: Function;
  normalize: Function;
  join: Function;
  dirname: Function;
  extname: Function;
  basename: Function;
  isAbsolute: Function;
}

// 继承TauriProcess
export interface TauriProcess {
  exit: Function;
  relaunch: Function;
}

// 继承TauriShell
export interface TauriShell {
  Command: any;
  Child: any;
  EventEmitter: any;
  open: Function;
}

// 继承TauriUpdate
export interface TauriUpdate {
  onUpdaterEvent: Function;
  installUpdate: Function;
  checkUpdate: Function;
}

// 继承TauriWindow
export interface TauriWindow {
  WebviewWindow: any;
  WebviewWindowHandle: any;
  WindowManager: any;
  CloseRequestedEvent: any;
  getCurrent: any;
  getAll: any;
  appWindow: any;
  LogicalSize: any;
  PhysicalSize: any;
  LogicalPosition: any;
  PhysicalPosition: any;
  UserAttentionType: any;
  currentMonitor: any;
  primaryMonitor: any;
  availableMonitors: any;
}

// 继承TauriOs
export interface TauriOs {
  platform: Function;
  version: Function;
  type: Function;
  arch: Function;
  tempdir: Function;
  locale: Function;
}